backend/requirements.txt
fastapi>=0.109.0
python-multipart>=0.0.9
uvicorn[standard]>=0.27.0
celery[redis]>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
cryptography>=42.0.0
